refactor: TraverseCultures recursive + extract TranslationReader class
