+++ src/kli.Localize.Generator/Internal/LocalizeCodeGeneratorCore.cs
@@ -1,29 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
-using Newtonsoft.Json;
-using Translations = System.Collections.Generic.Dictionary<string, string>;
 
 namespace kli.Localize.Generator.Internal
 {
     internal class LocalizeCodeGeneratorCore
     {
-        private readonly Action<Diagnostic> reportDiagnostic;
+        private readonly TranslationReader translationReader;
 
-        public LocalizeCodeGeneratorCore(Action<Diagnostic> reportDiagnostic)
+        public LocalizeCodeGeneratorCore(TranslationReader translationReader)
         {
-            this.reportDiagnostic = reportDiagnostic;
+            this.translationReader = translationReader;
         }
 
         public SourceText CreateClass(GeneratorDataContext context)
         {
-            var translations = this.ReadTranslations(context.OriginFilePath);
+            var translations = this.translationReader.Read(context.OriginFilePath);
 
             var classDeclaration = SyntaxFactory.ClassDeclaration(context.GeneratedClassName)
                 .AddModifiers(SyntaxFactory.Token(SyntaxKind.PublicKeyword))
@@ -95,18 +91,13 @@ namespace kli.Localize.Generator.Internal
 
                         private T TraverseCultures<T>(CultureInfo cultureInfo, SelectorFunc<T> selectorFunc)
                         {
-                            while (cultureInfo != CultureInfo.InvariantCulture)
+                            if (resources.TryGetValue(cultureInfo, out Translations translations))
                             {
-                                if (resources.TryGetValue(cultureInfo, out Translations translations))
-                                {
-                                    if (selectorFunc(translations, out T result))
-                                        return result;
-                                }
-                                cultureInfo = cultureInfo.Parent;
+                                if (selectorFunc(translations, out T result) || cultureInfo == CultureInfo.InvariantCulture)
+                                    return result;
                             }
 
-                            selectorFunc(resources[CultureInfo.InvariantCulture], out T retVal);
-                            return retVal;
+                            return TraverseCultures<T>(cultureInfo.Parent, selectorFunc);
                         }
                     }   
                 ").GetRoot();
@@ -118,7 +109,7 @@ namespace kli.Localize.Generator.Internal
 
         private MemberDeclarationSyntax GetTranslationDictionaryFieldDeclaration(CultureData ctx)
         {
-            var translations = this.ReadTranslations(ctx.FilePath);
+            var translations = this.translationReader.Read(ctx.FilePath);
             var entries = translations.Select(t => $"{{ \"{t.Key}\", {t.Value} }},");
             var source = $@"private static readonly Translations {ctx.Normalized} = new()
                             {{
@@ -137,51 +128,6 @@ namespace kli.Localize.Generator.Internal
                             }};";
 
             return SyntaxFactory.ParseMemberDeclaration(source);
-        }
-
-        private Translations ReadTranslations(string filePath)
-        {
-            try
-            {
-                var allTranslations = JsonConvert.DeserializeObject<Translations>(File.ReadAllText(filePath));
-                var validTranslations = allTranslations
-                               .Where(t => this.IsValidIdentifier(t.Key));
-
-                foreach (var invalid in allTranslations.Except(validTranslations))
-                {
-                    var match = File.ReadLines(filePath)
-                        .Select((line, index) => new { line, lineNumber = index })
-                        .FirstOrDefault(x => x.line.Contains(invalid.Key) && x.line.Contains(invalid.Value));
-
-                    this.ReportDiagnostic(filePath,
-                        $"Json property key must be a valid C# identifier: '{invalid.Key}'",
-                        match?.lineNumber ?? 0);
-                }
-
-                return validTranslations.ToDictionary(t => t.Key, t => JsonConvert.ToString(t.Value));
-            }
-            catch (JsonReaderException ex)
-            {
-                throw new JsonException($"Failed to parse json: '{filePath}'. Only string values are allowed. InnerException: '{ex.Message}'", ex);
-            }
-        }
-
-        private bool IsValidIdentifier(string identifier)
-        {
-            return SyntaxFacts.IsValidIdentifier(identifier)
-                && SyntaxFacts.GetKeywordKind(identifier) == SyntaxKind.None
-                && SyntaxFacts.GetContextualKeywordKind(identifier) == SyntaxKind.None;
-        }
-
-        private void ReportDiagnostic(string filePath, string message, int line = 0,
-            DiagnosticSeverity severity = DiagnosticSeverity.Warning)
-        {
-            var linePosition = new LinePosition(line, 0);
-            var diagnostic = Diagnostic.Create(
-                new DiagnosticDescriptor("SGL0001", "Source Generators", message, "Source Generators", severity, true),
-                Location.Create(filePath, new TextSpan(), new LinePositionSpan(linePosition, linePosition)));
-
-            this.reportDiagnostic(diagnostic);
         }
 
         private string CreateMemberHeader(string value) => $"///<summary>Similar to: {value}</summary>";
